requirements.txt
torch
pandas~=2.2.2
numpy~=1.26.4
pillow~=10.3.0
torchvision
scipy~=1.13.0
tqdm